Update README.md
Stranger Things is a critically acclaimed sci-fi horror series created by the Duffer Brothers. Premiering on Netflix in 2016, the show blends the charm of 80s nostalgia with a thrilling, supernatural plot. Set in the fictional town of Hawkins, Indiana, it follows a group of kids as they encounter strange, terrifying events linked to government experiments, parallel dimensions, and otherworldly creatures.